// src/interfaces/enums.ts
export enum AvailabilityStatus {
  AVAILABLE = "AVAILABLE",
  UNAVAILABLE = "UNAVAILABLE",
  RESERVATED = "RESERVATED",
}

export enum ReservationStatus {
  PENDING = "pending",
  CONFIRMED = "confirmed",
  CANCELLED = "cancelled",
}

export enum FlightCategory {
  LOCAL = "local",
  CROSS_COUNTRY = "cross_country",
  INSTRUCTION = "instruction",
  TOURISM = "tourism",
  TRAINING = "training",
  MAINTENANCE = "maintenance",
  PRIVATE = "private",
  CORPORATE = "corporate",
}